MAINT: Move f2pyarg helpers to service
